make new donut chart component and use everywhere
