+++ src/components/network/TokenDistributionChart.tsx
@@ -1,13 +1,13 @@
 /** @jsxImportSource @emotion/react */
 import { HTMLAttributes } from "react";
-import { css, useTheme } from "@emotion/react";
+import { css } from "@emotion/react";
 import LoadingSpinner from "../../assets/loading.gif";
-import Chart from "react-apexcharts";
 import { StatItem } from "./StatItem";
 import { formatNumber, nFormatter } from "../../utils/number";
 import { useTotalIssuance } from "../../hooks/useTotalIssuance";
 import { useDelegatedSupply } from "../../hooks/useDelegatedSupply";
 import { useAppStats } from "../../contexts";
+import { DonutChart } from "../DonutChart";
 
 const chartContainer = css`
   display: flex;
@@ -36,7 +36,6 @@ export type TokenDistributionChartProps = HTMLAttributes<HTMLDivElement>;
 export const TokenDistributionChart = () => {
 	const { state: { tokenLoading, tokenStats } } = useAppStats();
 	const token = tokenStats;
-	const theme = useTheme();
 	const totalIssuance = useTotalIssuance();
 	const delegated = useDelegatedSupply();
 
@@ -74,74 +73,22 @@ export const TokenDistributionChart = () => {
 					value={`${formatNumber(totalIssuance, { decimalPlaces: 2 })} 𝞃`}
 				/>
 			</div>
-			<Chart
+			<DonutChart
 				options={{
 					labels: [
 						`Circulating Delegated/Staked (${delegatedPercent}% of ${totalIssuanceFormatted})`,
 						`Circulating Free (${circulatingPercent}% of ${totalIssuanceFormatted})`,
 						`Unissued (${(
 							100 -
-              (totalIssuance.toNumber() / token.totalSupply) * 100
+							(totalIssuance.toNumber() / token.totalSupply) * 100
 						).toFixed(2)}% of ${nFormatter(token.totalSupply, 2)})`,
 					],
-					colors: ["#14dec2", "#FF9900", "#848484"],
-					dataLabels: {
-						enabled: false,
-						formatter: (text: string | number | number[], op: number) => {
-							return op.toFixed(2);
-						},
-					},
-					stroke: {
-						show: true,
-						curve: "smooth",
-						lineCap: "butt",
-						colors: [theme.palette.primary.dark],
-						width: 6,
-						dashArray: 0,
-					},
-					responsive: [
-						{
-							breakpoint: 767,
-							options: {
-								chart: {
-									height: 320,
-								},
-								stroke: {
-									width: 4,
-								},
-							},
-						},
-						{
-							breakpoint: 599,
-							options: {
-								chart: {
-									height: 270,
-								},
-								stroke: {
-									width: 2,
-								},
-							},
-						},
-					],
-					legend: {
-						show: true,
-						position: "bottom",
-						horizontalAlign: "left",
-						floating: false,
-						fontSize: "13px",
-						labels: {
-							colors: undefined,
-							useSeriesColors: true,
-						},
-					},
 				}}
 				series={[
 					token.delegatedSupply,
 					totalIssuance.toNumber() - token.delegatedSupply,
 					token.totalSupply - totalIssuance.toNumber(),
 				]}
-				type='donut'
-				height={400}
 			/>
 		</div>
 	);
